Time: 22 ms (84.23%), Space: 59.3 MB (38.86%) - LeetHub

+++ 2665-minimum-time-to-repair-cars/2665-minimum-time-to-repair-cars.cpp
@@ -4,7 +4,7 @@ public:
         for(int i=0;i<ranks.size();i++){
             if(cur>=ranks[i]){
             long long x=cur/ranks[i];
-            x=static_cast<long long>(sqrt(x));
+            x=(sqrt(x));
             n-=x;}
             if(n<=0)return 1;
         }
